refactor(ingestion-consumer): move commit submission into a commit manager
The sentinel check, the rdkafka commit, and the monitor that confirms
async commits landed move out of the consumer into CommitManager. The
consumer keeps the ledger, the settlement, and the poll bookkeeping, and
hands the manager the spans a poll is ready to commit.

No behaviour changes: one commit per completed poll, as before.
